<FEAT> Ajout d'exceptions dans UserService et PostService + Ajout de message d'erreur dans la classe MessageErreur + Ajout d'un log dans ExceptionManager + Ajout de throws dans les classes UserRestcontroller et PostRestController

+++ codecase_api/src/main/java/fr/swif/codecase_api/service/PostService.java
@@ -108,7 +108,10 @@ public class PostService {
    */
   // @Transactional surcharge le readOnly de la classe
   @Transactional
-  public void deletePost(int id) {
+  public void deletePost(int id) throws CodeCaseApiException{
+    if (!postRepository.existsById(id)) {
+      throw new CodeCaseApiException(MessagesErreur.POST_NOT_FOUND);
+    }
     postRepository.deleteById(id);
   }
 }
